+++ tests/test_utils.py
@@ -1,10 +1,15 @@
-from harvest_transformer.utils import get_file_from_url
+import pytest
+
+from harvest_transformer.utils import URLAccessError, get_file_from_url
 
 
 def test_get_file_from_url_fail():
     url = "https://invalid-test-url.com"
-    body = get_file_from_url(url)
-    assert body is None
+
+    with pytest.raises(URLAccessError) as e_info:
+        get_file_from_url(url)
+
+    assert "URL invalid" in str(e_info.value)
 
 
 def test_get_file_from_url_pass():